feat(partners): create partners page
